Move keyword enum to util package
