+++ src/main/java/duke/util/Keyword.java
@@ -1,4 +1,4 @@
-package duke.task;
+package duke.util;
 
 /**
  * Enum file for chatbot commands.
@@ -10,20 +10,22 @@ package duke.task;
 public enum Keyword {
     /** Exit message. */
     EXIT("bye"),
-    /** duke.task.Keyword for listing out all the task. */
+    /** duke.util.Keyword for listing out all the task. */
     LIST("list"),
-    /** duke.task.Keyword for marking task as done. */
+    /** duke.util.Keyword for marking task as done. */
     DONE("done"),
-    /** duke.task.Keyword for marking task as Deadline task." */
+    /** duke.util.Keyword for marking task as Deadline task." */
     DEADLINE("deadline", " /by ", "D"),
-    /** duke.task.Keyword for marking task as duke.task.Events. */
+    /** duke.util.Keyword for marking task as duke.task.Events. */
     EVENTS("event", " /at ", "E"),
-    /** duke.task.Keyword for marking task as Todos. */
+    /** duke.util.Keyword for marking task as Todos. */
     TODOS("todo", "", "T"),
-    /** duke.task.Keyword for deleting task. */
-    DELETE("delete");
+    /** duke.util.Keyword for deleting task. */
+    DELETE("delete"),
+    /** duke.util.Keyword for finding task. */
+    FIND("find");
 
-    /** duke.task.Keyword for command. */
+    /** duke.util.Keyword for command. */
     private String keyword;
     /** Separator for spliting up various input. */
     private String separator;
@@ -31,9 +33,9 @@ public enum Keyword {
     private  String saveWord;
 
     /**
-     * Constructor for duke.task.Keyword Enum.
+     * Constructor for duke.util.Keyword Enum.
      *
-     * @param keyword duke.task.Keyword for command
+     * @param keyword duke.util.Keyword for command
      */
     Keyword(String keyword) {
         this.keyword = keyword;
@@ -42,9 +44,9 @@ public enum Keyword {
     }
 
     /**
-     * Constructor for duke.task.Keyword Enum.
+     * Constructor for duke.util.Keyword Enum.
      *
-     * @param keyword duke.task.Keyword for command.
+     * @param keyword duke.util.Keyword for command.
      * @param separator Separator for spliting up various input.
      * @param saveWord saveWord for task.
      */
